# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: IronClaw Control-Plane API
  version: 0.1.0
  summary: Gateway change-approval, audit, and registry administration for a self-hosted IronClaw control-plane.
  description: |
    The control-plane exposes three surfaces:

    - **Gateway** (`/v1/changes`, `/v1/audit`) — submit capability changes, list
      what is pending, record human approve/reject decisions, and read the
      append-only audit log. Every control-plane mutation (persona, enabled
      tools, packages, wiring, permissions, mounts, create_agent) flows through
      here; there is no bypass path.
    - **Registry** (`/v1/registry/...`) — administer agent groups, messaging
      groups, wirings, users, roles, members, destinations, and read sessions.
      These endpoints return `503` until a registry backend is attached.
    - **Operational probes** (`/healthz`, `/readyz`, `/metrics`).

    All `POST`/`PUT` bodies are JSON. The Go JSON decoder is case-insensitive on
    keys; the property names below are the canonical forms the server **emits**.
  license:
    name: MIT
    url: https://github.com/nivardsec/ironclaw/blob/main/LICENSE
  contact:
    name: IronClaw
    url: https://github.com/nivardsec/ironclaw
...
servers:
  - url: http://127.0.0.1:8787
    description: Local development (`--dev`, loopback bind).
  - url: http://{tailnetHost}:{port}
    description: Production — bound to the control-plane's Tailscale interface address (no public port).
    variables:
...
      port:
...

...
security:
  - bearerAuth: []
...
paths:
...
  /v1/registry/agent-groups/{id}:
    put:
      tags: [registry]
      summary: Create or update an agent group.
      description: The path `id` is authoritative and overrides any `ID` in the body.
      operationId: putAgentGroup
      parameters:
        - $ref: "#/components/parameters/IDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentGroup"
      responses:
        "200":
          description: The stored agent group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentGroup"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
    get:
      tags: [registry]
      summary: Fetch an agent group by id.
      operationId: getAgentGroup
      parameters:
        - $ref: "#/components/parameters/IDPath"
      responses:
        "200":
          description: The agent group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentGroup"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
  /v1/registry/messaging-groups:
    post:
      tags: [registry]
      summary: Get or create a messaging group.
      description: Idempotent on the `(channelType, platformID, instance)` triple. An empty `unknownSenderPolicy` defaults to `strict`.
      operationId: createMessagingGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessagingGroupRequest"
      responses:
        "200":
          description: The resolved messaging group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagingGroup"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
  /v1/registry/messaging-groups/{id}:
    get:
      tags: [registry]
      summary: Fetch a messaging group by id.
      operationId: getMessagingGroup
      parameters:
        - $ref: "#/components/parameters/IDPath"
      responses:
        "200":
          description: The messaging group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagingGroup"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
  /v1/registry/messaging-groups/{id}/wirings:
    get:
      tags: [registry]
      summary: List the wirings for a messaging group.
      operationId: listWirings
      parameters:
        - $ref: "#/components/parameters/IDPath"
      responses:
        "200":
          description: The wirings (possibly empty).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wiring"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
  /v1/registry/wirings:
    post:
      tags: [registry]
      summary: Create or update a wiring.
      description: A server-assigned `ID` (prefixed `wr_`) is generated when omitted.
      operationId: createWiring
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Wiring"
      responses:
        "201":
          description: The stored wiring.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wiring"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
  /v1/registry/users/{id}:
    put:
      tags: [registry]
      summary: Create or update a user.
      description: The path `id` is authoritative and overrides any `ID` in the body.
      operationId: putUser
      parameters:
        - $ref: "#/components/parameters/IDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: The stored user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
    get:
      tags: [registry]
      summary: Fetch a user by id.
      operationId: getUser
      parameters:
        - $ref: "#/components/parameters/IDPath"
      responses:
        "200":
          description: The user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
  /v1/registry/roles:
    post:
      tags: [registry]
      summary: Grant a role to a user.
      operationId: grantRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Role"
      responses:
        "201":
          description: The granted role.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
  /v1/registry/roles/revoke:
    post:
      tags: [registry]
      summary: Revoke a role from a user.
      operationId: revokeRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Role"
      responses:
        "200":
          description: Role revoked.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    const: revoked
                required: [status]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
  /v1/registry/members:
    post:
      tags: [registry]
      summary: Add a user as a member of an agent group.
      operationId: addMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Member"
      responses:
        "201":
          description: The added member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
  /v1/registry/members/remove:
    post:
      tags: [registry]
      summary: Remove a member from an agent group.
      operationId: removeMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Member"
      responses:
        "200":
          description: Member removed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    const: removed
                required: [status]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
  /v1/registry/destinations:
    post:
      tags: [registry]
      summary: Allow an agent group to deliver to a destination.
      operationId: addDestination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationRequest"
      responses:
        "201":
          description: The added destination (echoes the request).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationRequest"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
  /v1/registry/destinations/check:
    get:
      tags: [registry]
      summary: Check whether a destination is allowed for an agent group.
      operationId: checkDestination
      parameters:
        - name: agentGroupID
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/AgentGroupID"
        - name: channelType
          in: query
          required: true
          schema:
            type: string
        - name: platformID
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Whether the destination is permitted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  allowed:
                    type: boolean
                required: [allowed]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
  /v1/registry/sessions:
    get:
      tags: [registry]
      summary: List all sessions.
      operationId: listSessions
      responses:
        "200":
          description: The sessions (possibly empty).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
  /v1/registry/sessions/{id}:
    get:
      tags: [registry]
      summary: Fetch a session by id.
      operationId: getSession
      parameters:
        - $ref: "#/components/parameters/IDPath"
      responses:
        "200":
          description: The session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
  /v1/registry/access:
    get:
      tags: [registry]
      summary: Check whether a user may access an agent group.
      operationId: checkAccess
      parameters:
        - name: userID
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/UserID"
        - name: agentGroupID
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/AgentGroupID"
      responses:
        "200":
          description: The access decision and a human-readable reason.
          content:
            application/json:
              schema:
                type: object
                properties:
                  allowed:
                    type: boolean
                  reason:
                    type: string
                required: [allowed, reason]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/RegistryUnavailable"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Static bearer token, supplied to the server via `IRONCLAW_API_TOKEN`
        (or `--token` for `ironctl`). It is **all-or-nothing** — there are no
        per-endpoint scopes; a valid token authorizes every non-probe endpoint.
        The token is compared in constant time. `/healthz` and `/readyz` are
        always exempt. When the server is started without a token the check is
        disabled and the mesh (Tailscale) boundary is the sole control.
  parameters:
    IDPath:
      name: id
      in: path
      required: true
      description: The resource id (authoritative for create/update).
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token (only when the server is started with a token).
      content:
        text/plain:
          schema:
            type: string
            example: unauthorized
    BadRequest:
      description: Malformed JSON body or invalid field values.
      content:
        text/plain:
          schema:
            type: string
    NotFound:
      description: Resource not found.
      content:
        text/plain:
          schema:
            type: string
    InternalError:
      description: Internal error reading from the backing store.
      content:
        text/plain:
          schema:
            type: string
...
    RegistryUnavailable:
      description: No registry backend is attached to this control-plane.
      content:
        text/plain:
          schema:
            type: string
            example: registry not configured
...
  schemas:
...
    UserID:
      type: string
      description: Platform identity, formatted `<channel>:<handle>`.
      examples: ["slack:U123", "telegram:42"]
    AgentGroupID:
      type: string
    MessagingGroupID:
      type: string
    SessionID:
      type: string

    # --- enums (see internal/contract/enums.go) ---
...
    UnknownSenderPolicy:
      type: string
      enum: [strict, request_approval, public]
      default: strict
    EngageMode:
      type: string
      enum: [pattern, mention, mention-sticky]
    SenderScope:
      type: string
      enum: [all, known]
    IgnoredMessagePolicy:
      type: string
      enum: [drop, accumulate]
    SessionMode:
      type: string
      enum: [shared, per-thread, agent-shared]
...
    AgentGroup:
      type: object
      description: A configured agent (workspace, persona, container config). Only host-facing identity fields are exposed here.
      properties:
        ID:
          $ref: "#/components/schemas/AgentGroupID"
        Name:
          type: string
        Folder:
          type: string
...
    MessagingGroup:
      type: object
      description: A single chat/channel on one platform, identified by the (ChannelType, PlatformID, Instance) triple.
      properties:
        ID:
          $ref: "#/components/schemas/MessagingGroupID"
        ChannelType:
          type: string
        PlatformID:
          type: string
        Instance:
          type: string
        IsGroup:
          type: boolean
        UnknownSenderPolicy:
          $ref: "#/components/schemas/UnknownSenderPolicy"
    MessagingGroupRequest:
      type: object
      description: Body for get-or-create of a messaging group (lower-case JSON keys).
      properties:
        channelType:
          type: string
        platformID:
          type: string
        instance:
          type: string
        isGroup:
          type: boolean
        unknownSenderPolicy:
          $ref: "#/components/schemas/UnknownSenderPolicy"
      required: [channelType, platformID]
    Wiring:
      type: object
      description: Links a messaging group to an agent group with engage/session policy.
      properties:
        ID:
          type: string
        MessagingGroupID:
          $ref: "#/components/schemas/MessagingGroupID"
        AgentGroupID:
          $ref: "#/components/schemas/AgentGroupID"
        EngageMode:
          $ref: "#/components/schemas/EngageMode"
        EngagePattern:
          type: string
        SenderScope:
          $ref: "#/components/schemas/SenderScope"
        IgnoredMessagePolicy:
          $ref: "#/components/schemas/IgnoredMessagePolicy"
        SessionMode:
          $ref: "#/components/schemas/SessionMode"
        Priority:
          type: integer
    User:
      type: object
      description: A platform identity (id is `<channel>:<handle>`).
      properties:
        ID:
          $ref: "#/components/schemas/UserID"
        Kind:
          type: string
        DisplayName:
          type: string
    Role:
      type: object
      description: Grants a user owner or admin privilege. A null `AgentGroupID` is a global role; otherwise it is scoped to that agent group.
      properties:
        UserID:
          $ref: "#/components/schemas/UserID"
        Role:
          type: string
          enum: [owner, admin]
        AgentGroupID:
          oneOf:
            - $ref: "#/components/schemas/AgentGroupID"
            - type: "null"
      required: [UserID, Role]
    Member:
      type: object
      description: Records unprivileged access of a user to an agent group.
      properties:
        UserID:
          $ref: "#/components/schemas/UserID"
        AgentGroupID:
          $ref: "#/components/schemas/AgentGroupID"
      required: [UserID, AgentGroupID]
    Session:
      type: object
      description: A per-(agent group, messaging group, thread) conversation with its own sandbox and queue pair.
      properties:
        ID:
          $ref: "#/components/schemas/SessionID"
        AgentGroupID:
          $ref: "#/components/schemas/AgentGroupID"
        MessagingGroupID:
          $ref: "#/components/schemas/MessagingGroupID"
        ThreadID:
          type: string
          nullable: true
        ContainerStatus:
          type: string
        LastActive:
          type: string
          format: date-time
    DestinationRequest:
      type: object
      description: Body to allow an agent group to deliver to a (channelType, platformID) destination.
      properties:
        agentGroupID:
          $ref: "#/components/schemas/AgentGroupID"
        channelType:
          type: string
        platformID:
          type: string
      required: [agentGroupID, channelType, platformID]